remove packs and extract custom js for boostrap tooltips
